+++ src/resources/Tasks.ts
@@ -1,3 +1,4 @@
+import { iso8601Duration } from "../duration";
 import Client from "../Client";
 import Task, { CreateTaskParams } from "../types/Task";
 
@@ -13,6 +14,10 @@ export default class Tasks {
       ...params,
       scheduled_for: params.scheduledFor?.toISOString() ?? params.scheduled_for,
       scheduledFor: undefined,
+      delay:
+        typeof params.delay === "object"
+          ? iso8601Duration(params.delay)
+          : params.delay,
     };
 
     const defaultParams = { queue: "default" };